Bug Fix:
Bug: If user presses settings icon n times then they must click
back button n times to get back to previous Fragment.

Cause: Every time settings icon was clicked a new PreferenceFragment
was added created and added to the backstack of Fragments.

Fix: When settings is clicked check that the PreferenceFragment is not
already being displayed, and if it is not then display it, otherwise
do nothing.